[FrontEnd] Thanh Toán

+++ Web_DongHo_WebAssembly/Models/BillDetailVM.cs
@@ -6,7 +6,7 @@
         public int BillId { get; set; }
         public int ProductId { get; set; }
         public int Quantity { get; set; }
-        public float UnitPrice { get; set; }
+        public decimal UnitPrice { get; set; }
         public float TotalPrice { get; set; }
     }
 }
